hw4/taxi_rides_ny/requirements.txt
pandas
pyarrow
google-cloud-storage
polars